2006-08-11  Phil Muldoon <pmuldoon@redhat.com>
	* src/jni/com_redhat_ftk_EventViewer.c: Revert to 1.7. Revert attempt to fix 64 breakage.
	Await orginal author input.
	*  src/widgets/ftkeventviewer.c: Revert to 1.30. Ditto comment.

+++ frysk-gtk/tlwidgets/src/widgets/ftkeventviewer.c
@@ -3027,7 +3027,7 @@ ftk_eventviewer_set_trace_linestyle (FtkEventViewer * eventviewer,
  		FtkTrace *trace = ftk_ev_trace(eventviewer, traceID);
  		if (ftk_trace_selected(trace)) 
  		{
- 			traces = g_list_append(traces, (gpointer)(gint) traceID);
+ 			traces = g_list_append(traces, (gpointer) traceID);
  		}
  	}
